Moved password-strength component to Tools
