+++ libs/angular/src/auth/components/register.component.ts
@@ -25,7 +25,7 @@ import {
   AllValidationErrors,
   FormValidationErrorsService,
 } from "../../platform/abstractions/form-validation-errors.service";
-import { PasswordColorText } from "../../shared/components/password-strength/password-strength.component";
+import { PasswordColorText } from "../../tools/password-strength/password-strength.component";
 import { InputsFieldMatch } from "../validators/inputs-field-match.validator";
 
 import { CaptchaProtectedComponent } from "./captcha-protected.component";
