+++ libs/angular/src/tools/password-strength/password-strength.component.html
